requirements.txt
Byaldi==0.0.4
numpy>=1.26.0
Pillow==10.4.0
pytesseract==0.3.13
streamlit==1.38.0
opencv-python-headless
